requirements.txt
altgraph==0.17.3
beautifulsoup4==4.12.2
bidict==0.22.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
dnspython==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.3
packaging==23.1
pathy==0.10.1
pefile==2023.2.7
preshed==3.0.8
pydantic==1.10.7
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
python-engineio==4.4.1
python-socketio==5.8.0
pywin32-ctypes==0.2.0
requests==2.29.0
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
threadpoolctl==3.1.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
websocket-client==1.5.1
